Add more navbar and page items
Downloaded the video from the original website showcasing one of the VR
games. Also put a placeholder on the logo part, though this might not be
even used.

Looking to style the rest of the page for now; leaving header alone
since I don't know how to style the cart part